Factor out key pushing/registration for V3 workspaces
